feat: CoinService 로직 구현
